Split ZarrOpenActions into ZarrOpener (fiji) and ZarrOpenActions (fiji-ui)
Move the backend selection and ImageJ/BDV opening core into a new
fiji-layer ZarrOpener with no fiji-ui dependencies. ZarrOpenActions keeps
the UI orchestration (settings dispatch, importer/viewer dialogs, script,
help) and composes a ZarrOpener, delegating the open methods. Add
ZarrOpenerTest and javadocs for both classes.
