+++ src/main/java/sc/fiji/ome/zarr/open/ZarrOpenActions.java
@@ -28,14 +28,10 @@
  */
 package sc.fiji.ome.zarr.open;
 
-import org.janelia.saalfeldlab.n5.N5Reader;
 import org.janelia.saalfeldlab.n5.bdv.N5ViewerCreator;
 import org.janelia.saalfeldlab.n5.ij.N5Importer;
-import org.janelia.saalfeldlab.n5.imglib2.N5Utils;
-import org.janelia.saalfeldlab.n5.universe.N5Factory;
 import org.scijava.Context;
 import org.scijava.prefs.PrefService;
-import org.scijava.ui.UIService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -46,34 +42,26 @@ import java.nio.file.Paths;
 import java.util.Collections;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.function.Supplier;
-
-import com.google.gson.JsonSyntaxException;
 
 import net.imglib2.img.Img;
-import net.imglib2.img.display.imagej.ImageJFunctions;
-import net.imglib2.util.Cast;
 
 import ij.IJ;
-import sc.fiji.ome.zarr.pyramid.fiji.PyramidalBdv;
-import sc.fiji.ome.zarr.pyramid.exceptions.MultiImageDatasetException;
-import sc.fiji.ome.zarr.pyramid.exceptions.NoMatchingResolutionException;
-import sc.fiji.ome.zarr.open.exceptions.NonExistingResolutionLevelException;
-import sc.fiji.ome.zarr.pyramid.exceptions.NotAMultiscaleImageException;
-import sc.fiji.ome.zarr.open.exceptions.NotASingleScaleImageException;
-import sc.fiji.ome.zarr.pyramid.fiji.PyramidalDataset;
-import sc.fiji.ome.zarr.pyramid.backend.PyramidBackend;
-import sc.fiji.ome.zarr.pyramid.backend.PyramidContents;
-import sc.fiji.ome.zarr.pyramid.backend.n5.N5PyramidBackend;
-import sc.fiji.ome.zarr.pyramid.backend.zarrjava.ZarrJavaPyramidBackend;
-import sc.fiji.ome.zarr.open.options.ZarrOpeningSettings;
 import sc.fiji.ome.zarr.open.options.ZarrOpenBehavior;
+import sc.fiji.ome.zarr.open.options.ZarrOpeningSettings;
 import sc.fiji.ome.zarr.open.options.ZarrReaderBackend;
+import sc.fiji.ome.zarr.pyramid.fiji.PyramidalDataset;
 import sc.fiji.ome.zarr.ui.DnDActionChooser;
-import sc.fiji.ome.zarr.pyramid.fiji.PyramidalService;
-import sc.fiji.ome.zarr.util.BdvUtils;
 import sc.fiji.ome.zarr.util.ScriptUtils;
 
+/**
+ * Fiji-ui orchestration of the OME-Zarr opening pipeline: it reads the user's
+ * {@link ZarrOpeningSettings}, dispatches to the chosen open behavior, and
+ * provides the UI-facing actions (N5 importer/viewer dialogs, preset script,
+ * help) wired by the {@link DnDActionChooser}.
+ * <p>
+ * The actual loading and ImageJ/BigDataViewer opening lives in
+ * {@link ZarrOpener}, so this class only adds the UI concerns on top.
+ */
 public class ZarrOpenActions
 {
 	private static final Logger logger = LoggerFactory.getLogger( MethodHandles.lookup().lookupClass() );
@@ -86,11 +74,7 @@ public class ZarrOpenActions
 
 	private final Consumer< String > errorHandler;
 
-	private final ZarrOpeningSettings settings;
-
-	private PyramidContents< ? > cachedContents;
-
-	private int preferredResolutionLevel;
+	private final ZarrOpener opener;
 
 	/**
 	 * Loads {@link ZarrOpeningSettings} from {@code context} and opens
@@ -120,23 +104,58 @@ public class ZarrOpenActions
 		}
 	}
 
+	/**
+	 * Actions for {@code inputUri} with default opening settings, reporting
+	 * failures via {@code IJ::error}.
+	 */
 	public ZarrOpenActions( final URI inputUri, final Context context )
 	{
 		this( inputUri, context, null, IJ::error );
 	}
 
+	/**
+	 * Actions for {@code inputUri} using the given opening {@code settings}
+	 * (reader backend and preferred resolution), reporting failures via
+	 * {@code IJ::error}. Pass {@code null} settings to use the defaults.
+	 */
 	public ZarrOpenActions( final URI inputUri, final Context context, final ZarrOpeningSettings settings )
 	{
 		this( inputUri, context, settings, IJ::error );
 	}
 
+	/**
+	 * Actions for {@code inputUri} using the given opening {@code settings} and an
+	 * explicit error sink. Pass {@code null} settings to use the defaults.
+	 *
+	 * @param errorHandler receives a user-facing message when opening fails
+	 */
 	ZarrOpenActions( final URI inputUri, final Context context, final ZarrOpeningSettings settings,
 			final Consumer< String > errorHandler )
 	{
 		this.inputUri = inputUri;
 		this.context = context;
-		this.settings = settings;
 		this.errorHandler = errorHandler;
+		this.opener = new ZarrOpener( inputUri, context, readerBackend( settings ), preferredMaxWidth( settings ), errorHandler );
+	}
+
+	/**
+	 * Reader backend from the settings, or the default when no settings are given.
+	 */
+	private static ZarrReaderBackend readerBackend( final ZarrOpeningSettings settings )
+	{
+		return settings == null ? ZarrOpeningSettings.DEFAULT_READER_BACKEND : settings.getReaderBackend();
+	}
+
+	/**
+	 * Preferred maximum width from the settings, or {@code null} (= highest
+	 * resolution) when no settings are given or the behavior is
+	 * {@link ZarrOpenBehavior#IMAGEJ_HIGHEST_RESOLUTION}.
+	 */
+	private static Integer preferredMaxWidth( final ZarrOpeningSettings settings )
+	{
+		if ( settings == null || settings.getOpenBehavior() == ZarrOpenBehavior.IMAGEJ_HIGHEST_RESOLUTION )
+			return null;
+		return settings.getPreferredMaxWidth();
 	}
 
 	/**
@@ -173,216 +192,52 @@ public class ZarrOpenActions
 			logger.info( "Opened Zarr/N5 viewer with location: {}.", inputUri );
 	}
 
+	/**
+	 * Opens the dataset in ImageJ at the resolution selected by the settings.
+	 * Delegates to {@link ZarrOpener#openIJWithImage()}.
+	 */
 	public Object openIJWithImage()
 	{
-		try
-		{
-			return openPyramidImage(
-					() -> {
-						final PyramidContents< ? > contents = getContents();
-						final PyramidalDataset dataset = new PyramidalDataset( context, contents, preferredResolutionLevel );
-						context.getService( UIService.class ).show( dataset );
-						logger.info( "Opened dataset in ImageJ: {}", inputUri );
-						return null;
-					},
-					singleScaleImage -> ImageJFunctions.show( Cast.unchecked( singleScaleImage ) ) );
-		}
-		catch ( NoMatchingResolutionException e )
-		{
-			showNonMatchingResolutionError( e );
-		}
-		return null;
+		return opener.openIJWithImage();
 	}
 
 	/**
-	 * Opens the resolution level at {@code resolutionLevel} index as a new ImageJ dataset.
-	 * Index 0 is the highest resolution; each increment is the next coarser level.
-	 * <p>
-	 * Multiple calls — whether at the same or different level indices — each produce a separate
-	 * ImageJ {@code Dataset} (and a separate window), but all of them are backed by the same
-	 * {@link sc.fiji.ome.zarr.pyramid.backend.PyramidContents} object: the cached cell images
-	 * and volatile images are the single source of truth and are never loaded more than once
-	 * per resolution level.
-	 *
-	 * @param resolutionLevel 0-based index into the resolution pyramid
+	 * Opens the given resolution level of the dataset in ImageJ (0 = highest
+	 * resolution). Delegates to {@link ZarrOpener#openIJWithImage(int)}.
 	 */
 	public Object openIJWithImage( final int resolutionLevel )
 	{
-		try
-		{
-			return openPyramidImage(
-					() -> {
-						final PyramidContents< ? > contents = getContents();
-						if ( resolutionLevel < 0 || resolutionLevel >= contents.numResolutionLevels() )
-							throw new NonExistingResolutionLevelException( resolutionLevel, contents.numResolutionLevels() );
-						final PyramidalDataset dataset = new PyramidalDataset( context, contents, resolutionLevel );
-						context.getService( UIService.class ).show( dataset );
-						logger.info( "Opened dataset at resolution level {} in ImageJ: {}", resolutionLevel, inputUri );
-						return null;
-					},
-					singleScaleImage -> ImageJFunctions.show( Cast.unchecked( singleScaleImage ) ) );
-		}
-		catch ( NonExistingResolutionLevelException e )
-		{
-			showNonExistingResolutionLevelError( e );
-		}
-		return null;
+		return opener.openIJWithImage( resolutionLevel );
 	}
 
+	/**
+	 * Opens the dataset in BigDataViewer. Delegates to
+	 * {@link ZarrOpener#openBDVWithImage()}.
+	 */
 	public Object openBDVWithImage()
 	{
-		try
-		{
-			return openPyramidImage(
-					() -> {
-						final PyramidalBdv< ? > dataset = new PyramidalBdv<>( context, getContents() );
-						final PyramidalService pyramidalService = context.getService( PyramidalService.class );
-						final Object result = BdvUtils.showBdvAndRegisterDataset( dataset, pyramidalService );
-						logger.info( "Opened dataset in BigDataViewer: {}", inputUri );
-						return result;
-					},
-					singleScaleImage -> null );
-		}
-		catch ( NoMatchingResolutionException e )
-		{
-			showNonMatchingResolutionError( e );
-		}
-		return null;
-	}
-
-	private PyramidContents< ? > getContents()
-	{
-		if ( cachedContents == null )
-		{
-			final Integer preferredWidth;
-			if ( settings == null || settings.getOpenBehavior().equals( ZarrOpenBehavior.IMAGEJ_HIGHEST_RESOLUTION ) )
-				preferredWidth = null;
-			else
-				preferredWidth = settings.getPreferredMaxWidth();
-
-			final ZarrReaderBackend backend = settings == null
-					? ZarrOpeningSettings.DEFAULT_READER_BACKEND
-					: settings.getReaderBackend();
-			final PyramidBackend pyramidBackend = ( backend == ZarrReaderBackend.ZARR_JAVA )
-					? new ZarrJavaPyramidBackend()
-					: new N5PyramidBackend();
-
-			final PyramidContents< ? > contents = pyramidBackend.load( inputUri );
-			preferredResolutionLevel = contents.selectResolutionLevel( preferredWidth );
-			cachedContents = contents;
-		}
-		return cachedContents;
-	}
-
-	private Object openPyramidImage( final Supplier< Object > multiScaleOpener, final Function< Img< ? >, Object > singleScaleOpener )
-	{
-		try
-		{
-			return multiScaleOpener.get();
-		}
-		catch ( MultiImageDatasetException e )
-		{
-			showMultiImageNotSupported( e );
-		}
-		catch ( NotAMultiscaleImageException e )
-		{
-			logger.warn( "Not a multiscale image: {}", e.getMessage() );
-			showSingleScaleNotSupported();
-			// TODO: openSingleScaleImage( singleScaleOpener ) when single-scale support is added
-		}
-		catch ( IllegalArgumentException | JsonSyntaxException e )
-		{
-			showNonZarrError( e );
-		}
-		return null;
-	}
-
-	private void showMultiImageNotSupported( final MultiImageDatasetException e )
-	{
-		errorHandler.accept( e.getMessage() );
-		logger.info( e.getMessage() );
-	}
-
-	private void showSingleScaleNotSupported()
-	{
-		errorHandler.accept(
-				"Opening a single resolution OME-Zarr dataset, as was found in: " + inputUri + ", is currently not supported.\n\n"
-						+ "Consider opening one level higher in the hierarchy instead." );
-		logger.info( "Opening a single resolution OME-Zarr dataset, as was found in: {}, is currently not supported.", inputUri );
-	}
-
-	private void showSingleScaleError( final Exception e )
-	{
-		errorHandler.accept( "Could not open dataset as image: " + inputUri + "\n\n"
-				+ "Consider opening one level higher or lower in the hierarchy instead." );
-		logger.warn( "Could not open dataset as single resolution image: {}. Error message: {}", inputUri, e.getMessage() );
-	}
-
-	private void showNonZarrError( final Exception e )
-	{
-		errorHandler.accept( "Could not open dataset as image: " + inputUri + "\n\n"
-				+ "The opener for OME-Zarr only supports locations that contains OME-Zarr metadata, i.e. .zattrs, .zgroup, or zarr.json files." );
-		logger.warn( "Could not open dataset image: {}. Error message: {}", inputUri, e.getMessage() );
-	}
-
-	private void showNonMatchingResolutionError( final Exception e )
-	{
-		errorHandler.accept( "Safety check failed when opening dataset: " + inputUri + "\n\r\n" + e.getMessage() + "\n\r\n"
-				+ "If the image size is okay for this computer, please adjust the setting in\nPlugins > OME-Zarr > Settings > Opening Behavior Settings to still open the image." );
-		logger.warn( "Not opening dataset: {}. Error message: {}", inputUri, e.getMessage() );
-	}
-
-	private void showNonExistingResolutionLevelError( final NonExistingResolutionLevelException e )
-	{
-		errorHandler.accept( "Could not open resolution level from dataset: " + inputUri + "\n\n" + e.getMessage() );
-		logger.warn( "Could not open resolution level: {}. Error message: {}", inputUri, e.getMessage() );
+		return opener.openBDVWithImage();
 	}
 
 	Object openImage( final Function< PyramidalDataset, Object > multiScaleImageOpener,
 			final Function< Img< ? >, Object > singleScaleImageOpener )
 	{
-		try
-		{
-			return openPyramidImage(
-					() -> {
-						final PyramidContents< ? > contents = getContents();
-						final PyramidalDataset dataset = new PyramidalDataset( context, contents, preferredResolutionLevel );
-						final Object result = multiScaleImageOpener.apply( dataset );
-						logger.info( "Opened dataset: {}", inputUri );
-						return result;
-					},
-					singleScaleImageOpener );
-		}
-		catch ( NoMatchingResolutionException e )
-		{
-			showNonMatchingResolutionError( e );
-		}
-		return null;
+		return opener.openImage( multiScaleImageOpener, singleScaleImageOpener );
 	}
 
-	private Object openSingleScaleImage( final Function< Img< ? >, Object > singleScaleImageOpener ) throws NotASingleScaleImageException
-	{
-		N5Reader reader = new N5Factory().openReader( inputUri.toString() );
-		Img< ? > img;
-		try
-		{
-			img = N5Utils.open( reader, "" );
-		}
-		catch ( Exception e )
-		{
-			throw new NotASingleScaleImageException( inputUri.toString(), e );
-		}
-		Object result = singleScaleImageOpener.apply( img );
-		logger.info( "Opened single scale image: {}", inputUri );
-		return result;
-	}
-
+	/**
+	 * Opens the Fiji script editor pre-filled with a scriptlet that opens the
+	 * dataset, so the user can adapt it for macros/scripts.
+	 */
 	public void runScript()
 	{
 		logger.info( "Attempt to execute script on location: {}.", inputUri );
 		ScriptUtils.executePresetScript( context, inputUri, errorHandler );
 	}
 
+	/**
+	 * Opens the project's help page in the system web browser.
+	 */
 	public void showHelp()
 	{
 		try
